+++ unpack/BFLIM.py
@@ -26,6 +26,42 @@ L4 = 0x0c
 A4 = 0x0d
 ETC1_2 = 0x13
 
+PIXEL_SIZES = {
+	L8: 1,
+	A8: 1,
+	LA4: 1,
+	LA8: 2,
+	HILO8: 2,
+	RGB565: 2,
+	RGB8: 3,
+	RGBA5551: 2,
+	RGBA4: 2,
+	RGBA8: 4,
+	ETC1: None,
+	ETC1A4: None,
+	L4: 0.5,
+	A4: 0.5,
+	ETC1_2: None
+}
+
+FORMAT_NAMES = {
+	L8: 'L8',
+	A8: 'A8',
+	LA4: 'LA4',
+	LA8: 'LA8',
+	HILO8: 'HILO8',
+	RGB565: 'RGB565',
+	RGB8: 'RGB8',
+	RGBA5551: 'RGBA5551',
+	RGBA4: 'RGBA4',
+	RGBA8: 'RGBA8',
+	ETC1: 'ETC1',
+	ETC1A4: 'ETC1A4',
+	L4: 'L4',
+	A4: 'A4',
+	ETC1_2: 'ETC1',
+}
+
 ETC1_MODIFIERS = [
 	[2, 8],
 	[5, 17],
@@ -38,7 +74,7 @@ ETC1_MODIFIERS = [
 ]
 
 class extractBFLIM(ClsFunc, rawutil.TypeReader):
-	def main(self, filename, data):
+	def main(self, filename, data, opts={}):
 		self.outfile = make_outfile(filename, 'png')
 		data = self.readheader(data)
 		if self.format in (ETC1, ETC1A4, ETC1_2):
@@ -69,44 +105,69 @@ class extractBFLIM(ClsFunc, rawutil.TypeReader):
 		hdata = self.unpack(BFLIM_IMAG_HDR_STRUCT, hdr)
 		if hdata[0] != b'imag':
 			error('Invalid magic for imag header: %s' % hdata[0])
-		#parsesize = hdata[1] #0x10
+		#headerlen = hdata[1] #0x10
 		self.width = hdata[2]
 		self.height = hdata[3]
 		self.align = hdata[4]
 		self.format = hdata[5]
+		print('Color format: %s' % FORMAT_NAMES[self.format])
+		self.pxsize = PIXEL_SIZES[self.format]
 		self.swizzle = hdata[6]
+		print('Texture swizzling: %d' % self.swizzle)
 		#datalen = hdata[7]
 		
 	def extract(self, data):
 		img = Image.new('RGBA', (self.width, self.height))
-		pixels = img.load()
-		dataw = 1 << int(math.ceil(math.log(self.width, 2)))
-		datah = 1 << int(math.ceil(math.log(self.height, 2)))
-		tiles_width = math.ceil(self.width / 8)
-		tiles_height = math.ceil(self.height / 8)
-		for tiley in range(0, tiles_height):
-			for tilex in range(0, tiles_width):
-				for y1 in range(0, 2):
-					for x1 in range(0, 2):
-						for y2 in range(0, 2):
-							for x2 in range(0, 2):
-								for y3 in range(0, 2):
-									for x3 in range(0, 2):
-										pixel_x = (x3 + (x2 * 2) + (x1 * 4) + (tilex * 8))
-										pixel_y = (y3 + (y2 * 2) + (y1 * 4) + (tiley * 8))
-										if pixel_x >= self.width:
-											continue
-										if pixel_y >= self.height:
-											continue
-										data_x = (x3 + (x2 * 4) + (x1 * 16) + (tilex * 64))
-										data_y = ((y3 * 2) + (y2 * 8) + (y1 * 32) + (tiley * dataw * 8))
-										pos = data_x + data_y
-										px = self.getpixel(data, pos)
-										pixels[pixel_x, pixel_y] = px
+		self.pixels = img.load()
+		
+		datawidth = 1 << int(math.ceil(math.log(self.width, 2)))
+		dataheight = 1 << int(math.ceil(math.log(self.height, 2)))
+		
+		tiles_x = math.ceil(self.width / 8)
+		tiles_y = math.ceil(self.height / 8)
+		
+		#8x8px tiles
+		for ytile in range(tiles_y):
+			for xtile in range(tiles_x):
+				tilepos = int(((ytile * 64 * tiles_x) + (xtile * 64)) * self.pxsize)
+				tile = data[tilepos:tilepos + 64]
+				self.extract_tile(tile, xtile, ytile)
+				
 		img = self.deswizzle(img)
 		img.save(self.outfile, 'PNG')
 		
-	def getpixel(self, data, ptr):
+	def extract_tile(self, tile, xtile, ytile):
+		for ysub in range(2):
+			for xsub in range(2):
+				subpos = int(((ysub * 32) + (xsub * 16)) * self.pxsize)
+				sub = tile[subpos:subpos + 16]
+				for ygroup in range(2):
+					for xgroup in range(2):
+						grppos = int(((ygroup * 8) + (xgroup * 4)) * self.pxsize)
+						grp = sub[grppos:grppos + 4]
+						if self.pxsize == 0.5:
+							for ypix in range(2):
+								xpix = grp[ypix:ypix + 1]
+								px1 = self.getpixel(xpix, subpx=1)
+								px2 = self.getpixel(xpix, subpx=0)
+								outpos_y = (ytile * 8) + (ysub * 4) + (ygroup * 2) + ypix
+								outpos_x = (xtile * 8) + (xsub * 4) + (xgroup * 2)
+								if outpos_x >= self.width or outpos_y >= self.height:
+									continue
+								self.pixels[outpos_x, outpos_y] = px1
+								self.pixels[outpos_x + 1, outpos_y] = px2
+						else:
+							for ypix in range(2):
+								for xpix in range(2):
+									pixpos = ((ypix * 2) + xpix) * self.pxsize
+									pixel = grp[pixpos:pixpos + self.pxsize]
+									rgba = self.getpixel(pixel)
+									outpos_y = (ytile * 8) + (ysub * 4) + (ygroup * 2) + ypix
+									outpos_x = (xtile * 8) + (xsub * 4) + (xgroup * 2) + xpix
+									self.pixels[outpos_x, outpos_y] = rgba
+									#assert rgba[3] == 0
+									
+	def getpixel(self, data, ptr=0, subpx=0):
 		if self.format == L8:
 			l = self.unpack_from('B', data, ptr)[0]
 			px = (l, l, l, 255)
@@ -114,7 +175,7 @@ class extractBFLIM(ClsFunc, rawutil.TypeReader):
 			px = (0, 0, 0) + self.unpack_from('B', data, ptr)
 		elif self.format == LA4:
 			la = self.unpack_from('B', data, ptr)[0]
-			l = (la >> 4) *0x11
+			l = (la >> 4) * 0x11
 			a = (la & 0x0f) * 0x11
 			px = (l, l, l, a)
 		elif self.format == LA8:
@@ -145,8 +206,8 @@ class extractBFLIM(ClsFunc, rawutil.TypeReader):
 			a = (ba & 0x01) * 0x11
 			px = (r, g, b, a)
 		elif self.format == L4:
-			val = self.unpack_from('B', data, ptr // 2)[0]
-			shift = (ptr & 1) * 4
+			val = self.unpack_from('B', data, ptr)[0]
+			shift = subpx * 4
 			l = ((val >> shift) & 0x0f) * 0x11
 			px = (l, l, l, 255)
 		elif self.format == A4:
@@ -165,21 +226,22 @@ class extractBFLIM(ClsFunc, rawutil.TypeReader):
 			img = img.rotate(90)
 			img = img.transpose(Image.FLIP_LEFT_RIGHT)
 		return img
-	
+		
 	def diff_complement(self, val, bits):
 		if val >> (bits - 1) == 0:
 			return val
 		return val - (1 << bits)
 		
 	def decompress_ETC1(self, data):
+		#Based on ObsidianX's BFLIM ETC1 decompression algorithm (https://github.com/ObsidianX/3dstools)
 		has_alpha = (self.format == ETC1A4)
 		blklen = (16 if has_alpha else 8)
-		img = Image.new('RGBA', (self.width, self.height))
+		img = Image.new('RGBA', (self.height, self.width))
 		pixels = img.load()
-		tile_w = int(math.ceil(self.width / 8))
-		tile_h = int(math.ceil(self.height / 8))
-		tile_w = 1 << int(math.ceil(math.log(tile_w, 2)))
-		tile_h = 1 << int(math.ceil(math.log(tile_h, 2)))
+		tile_h = math.ceil(self.width / 8)
+		tile_w = math.ceil(self.height / 8)
+		tile_h = 1 << math.ceil(math.log(tile_h, 2))
+		tile_w = 1 << math.ceil(math.log(tile_w, 2))
 		ptr = 0
 		for tiley in range(0, tile_h):
 			for tilex in range(0, tile_w):
@@ -191,8 +253,8 @@ class extractBFLIM(ClsFunc, rawutil.TypeReader):
 							alphas = self.unpack_from('Q', block, 0)[0]
 							block = block[8:]
 						else:
-							alphas = (2 ** 64) - 1  #0xffff... for 8B
-						pxs = self.unpack('Q', block, 0)[0]
+							alphas = (2 ** 64) - 1  #0xffff... on 8B
+						pxs = self.unpack('Q', block)[0]
 						differential = (pxs >> 33) & 1
 						horizontal = (pxs >> 32) & 1
 						table1 = ETC1_MODIFIERS[(pxs >> 37) & 7]
@@ -201,10 +263,10 @@ class extractBFLIM(ClsFunc, rawutil.TypeReader):
 							r = (pxs >> 59) & 0x1f
 							g = (pxs >> 51) & 0x1f
 							b = (pxs >> 43) & 0x1f
-							r = (r << 3) | ((r >> 2) & 7)
-							g = (g << 3) | ((g >> 2) & 7)
-							b = (b << 3) | ((b >> 2) & 7)
-							color1 = (r, g, b)
+							r1 = (r << 3) | ((r >> 2) & 7)
+							g1 = (g << 3) | ((g >> 2) & 7)
+							b1 = (b << 3) | ((b >> 2) & 7)
+							color1 = (r1, g1, b1)
 							
 							r += self.diff_complement((pxs >> 56) & 7, 3)
 							g += self.diff_complement((pxs >> 48) & 7, 3)
@@ -218,19 +280,20 @@ class extractBFLIM(ClsFunc, rawutil.TypeReader):
 							g = ((pxs >> 52) & 0x0f) * 0x11
 							b = ((pxs >> 44) & 0x0f) * 0x11
 							color1 = (r, g, b)
+							
 							r = ((pxs >> 56) & 0x0f) * 0x11
 							g = ((pxs >> 48) & 0x0f) * 0x11
 							b = ((pxs >> 40) & 0x0f) * 0x11
 							color2 = (r, g, b)
 						amounts = pxs & 0xffff
-						signs = pxs & 0xffff
+						signs = (pxs >> 16) & 0xffff
 						for pixely in range(0, 4):
 							for pixelx in range(0, 4):
 								x = pixelx + (blockx * 4) + (tilex * 8)
 								y = pixely + (blocky * 4) + (tiley * 8)
-								if x >= self.width:
+								if x >= self.height:
 									continue
-								if y >= self.height:
+								if y >= self.width:
 									continue
 								offset = (pixelx * 4) + pixely
 								if horizontal: